Fix lint in datalake utils
