+++ ingestion/src/metadata/ingestion/source/database/datalake/metadata.py
@@ -59,13 +59,11 @@ from metadata.ingestion.source.database.datalake.models import (
     DatalakeTableSchemaWrapper,
 )
 from metadata.utils import fqn
-from metadata.utils.constants import DEFAULT_DATABASE
 from metadata.utils.datalake.datalake_utils import (
-    COMPLEX_COLUMN_SEPARATOR,
-    SUPPORTED_TYPES,
     clean_dataframe,
     fetch_dataframe,
 )
+from metadata.utils.constants import COMPLEX_COLUMN_SEPARATOR, DEFAULT_DATABASE
 from metadata.utils.filters import filter_by_schema, filter_by_table
 from metadata.utils.logger import ingestion_logger
