+++ ingestion/src/metadata/utils/constants.py
@@ -19,6 +19,7 @@ UTF_8 = "utf-8"
 CHUNKSIZE = 200000
 DEFAULT_DATABASE = "default"
 BUILDER_PASSWORD_ATTR = "password"
+COMPLEX_COLUMN_SEPARATOR = "_##"
 
 ES_SOURCE_TO_ES_OBJ_ARGS = {
     "caCerts": "ca_certs",
