+++ ingestion/src/metadata/utils/datalake/datalake_utils.py
@@ -16,43 +16,38 @@ from different auths and different file systems.
 
 
 from enum import Enum
-from typing import Any, Dict
 
 from metadata.ingestion.source.database.datalake.models import (
     DatalakeTableSchemaWrapper,
 )
-from metadata.utils.constants import CHUNKSIZE
+from metadata.utils.datalake.avro_dispatch import read_avro_dispatch
+from metadata.utils.datalake.csv_tsv_dispatch import (
+    read_csv_dispatch,
+    read_tsv_dispatch,
+)
+from metadata.utils.datalake.json_dispatch import read_json_dispatch
+from metadata.utils.datalake.parquet_dispatch import read_parquet_dispatch
 from metadata.utils.logger import utils_logger
 
 logger = utils_logger()
-COMPLEX_COLUMN_SEPARATOR = "_##"
-AZURE_PATH = "abfs://{bucket_name}@{account_name}.dfs.core.windows.net/{key}"
 logger = utils_logger()
 
 
-class FILE_FORMAT_DISPATCH_MAP:
+class FileFormatDispatchMap:
     @classmethod
     def fetch_dispatch(cls):
-        from metadata.utils.datalake.avro_dispatch import read_avro_dispatch
-        from metadata.utils.datalake.csv_tsv_dispatch import (
-            read_csv_dispatch,
-            read_tsv_dispatch,
-        )
-        from metadata.utils.datalake.json_dispatch import read_json_dispatch
-        from metadata.utils.datalake.parquet_dispatch import read_parquet_dispatch
-
         return {
-            SUPPORTED_TYPES.CSV: read_csv_dispatch,
-            SUPPORTED_TYPES.TSV: read_tsv_dispatch,
-            SUPPORTED_TYPES.AVRO: read_avro_dispatch,
-            SUPPORTED_TYPES.PARQUET: read_parquet_dispatch,
-            SUPPORTED_TYPES.JSON: read_json_dispatch,
-            SUPPORTED_TYPES.JSONGZ: read_json_dispatch,
-            SUPPORTED_TYPES.JSONZIP: read_json_dispatch,
+            SupportedTypes.CSV: read_csv_dispatch,
+            SupportedTypes.TSV: read_tsv_dispatch,
+            SupportedTypes.AVRO: read_avro_dispatch,
+            SupportedTypes.PARQUET: read_parquet_dispatch,
+            SupportedTypes.JSON: read_json_dispatch,
+            SupportedTypes.JSONGZ: read_json_dispatch,
+            SupportedTypes.JSONZIP: read_json_dispatch,
         }
 
 
-class SUPPORTED_TYPES(Enum):
+class SupportedTypes(Enum):
     CSV = "csv"
     TSV = "tsv"
     AVRO = "avro"
@@ -63,27 +58,7 @@ class SUPPORTED_TYPES(Enum):
 
     @property
     def return_dispatch(self):
-        return FILE_FORMAT_DISPATCH_MAP.fetch_dispatch().get(self)
-
-
-def return_azure_storage_options(config_source: Any) -> Dict:
-    connection_args = config_source.securityConfig
-    return {
-        "tenant_id": connection_args.tenantId,
-        "client_id": connection_args.clientId,
-        "account_name": connection_args.accountName,
-        "client_secret": connection_args.clientSecret.get_secret_value(),
-    }
-
-
-def dataframe_to_chunks(df):
-    """
-    Reads the Dataframe and returns list of dataframes broken down in chunks
-    """
-    return [
-        df[range_iter : range_iter + CHUNKSIZE]
-        for range_iter in range(0, len(df), CHUNKSIZE)
-    ]
+        return FileFormatDispatchMap.fetch_dispatch().get(self)
 
 
 def fetch_dataframe(
@@ -97,7 +72,7 @@ def fetch_dataframe(
     bucket_name: str = file_fqn.bucket_name
 
     try:
-        for supported_types_enum in SUPPORTED_TYPES:
+        for supported_types_enum in SupportedTypes:
             if key.endswith(supported_types_enum.value):
                 return supported_types_enum.return_dispatch(
                     config_source,
